fix for failing delete 6 days old articles job

+++ app/job_logic/job_logics.py
@@ -24,6 +24,8 @@ def run_add_articles_and_clusters_job() -> dict[str, object]:
 
 def run_delete_6_days_old_call_job(db):
     articles_ids_to_delete = query.get_articles_ids_older_than_6_days(db)
+    if not articles_ids_to_delete:
+        return 0
     vector_store.delete_embeddings_by_ids([str(art_id) for art_id in articles_ids_to_delete])
     query.delete_info_clusters_articles_with_ids(articles_ids_to_delete, db)
     db.commit()
